fixes #360
