+++ src/dishka/container.py
@@ -11,10 +11,10 @@ from dishka.entities.component import DEFAULT_COMPONENT, Component
 from dishka.entities.factory_type import FactoryType
 from dishka.entities.key import DependencyKey
 from dishka.entities.scope import BaseScope, Scope
-from ._adaptix.type_tools.basic_utils import is_type_alias_type
 from .container_objects import Exit
 from .context_proxy import ContextProxy
 from .dependency_source import Factory
+from .entities.type_alias_type import is_type_alias_type
 from .entities.validation_settigs import DEFAULT_VALIDATION, ValidationSettings
 from .exceptions import (
     ChildScopeNotFoundError,
@@ -158,7 +158,8 @@ class Container:
     ) -> Any:
         lock = self.lock
         if is_type_alias_type(dependency_type):
-            dependency_type = dependency_type.__value__
+            while hasattr(dependency_type, "__value__"):
+                dependency_type = dependency_type.__value__
         key = DependencyKey(dependency_type, component)
         try:
             if not lock:
