+++ tests/unit/type_alias_type_provider.py
@@ -1,0 +1,5 @@
+type Integer = int
+type Integer2 = int
+type ListFloat = list[float]
+type DictIntStr = dict[int, str]
+type WrappedInteger = Integer
